backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pysyslogclient==0.2.0
requests==2.31.0
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
cryptography==41.0.5
gunicorn==21.2.0
python-dateutil==2.8.2
email-validator==2.1.0.post1
